v17.0.5
- UI and performance improvements
- Categories 'Movies' and 'Series' were merged
- Allow to choose active channel grid, between:
    1. Default - Provided by the app.
    2. Lists - Extracted from the active IPTV lists.
    3. EPG - Extracted from the EPG.

+++ www/nodejs-project/modules/streamer/adapters/ts.js
@@ -11,26 +11,26 @@ class StreamerAdapterTS extends StreamerAdapterBase {
 	}
 	start(){
 		return new Promise((resolve, reject) => {
-			this.setCallback(success => {
-				if(success){
-					resolve()
-				} else {
-					reject()
-				}
-			})
+			this.setCallback(success => (success ? resolve : reject)())
 			const args = [this.url, this.opts]
 			if(global.config.get('ts-packet-filter-policy') == -1){
 				this.source = new Downloader(...args)
 			} else {
 				const useWorker = false
 				if(useWorker) {
-					const MultiWorker = require('../../multi-worker')
-					this.worker = new MultiWorker()
-					this.source = this.worker.load(path.join(__dirname, '../utils/joiner-worker'))
-					this.once('destroy', () => {
-						this.source && this.source.terminate()
-						this.worker && this.worker.terminate()
-					})
+					const exclusiveWorker = true
+					if(exclusiveWorker) {
+						const MultiWorker = require('../../multi-worker')
+						this.worker = new MultiWorker()
+						this.source = this.worker.load(path.join(__dirname, '../utils/joiner-worker'))
+						this.once('destroy', () => {
+							this.source && this.source.terminate()
+							this.worker && this.worker.terminate()
+						})
+					} else {
+						this.source = global.workers.load(path.join(__dirname, '../utils/joiner-worker'))
+						this.once('destroy', () => this.source && this.source.terminate())
+					}
 				} else {
 					this.source = new Joiner(...args)
 				}
